requirements.txt
Cython==0.29.24
pyenet==1.3.14
